feat: convert to v2 service
