--- src/services/hubspot-services/index.js
@@ -1,3 +1,0 @@
-import { hubspotFormSubmitService } from './hubspot-form-submit-service'
-
-export { hubspotFormSubmitService }
