+++ src/services/v2/hubspot/index.js
@@ -1,0 +1,1 @@
+export { HubspotService, hubspotService } from './hubspot-service'
